added observer for published promotions

+++ src/fr/justin/pannier/ArticleComparator.java
@@ -5,9 +5,9 @@ import java.util.Comparator;
 public class ArticleComparator implements Comparator<Article> {
     @Override
     public int compare(Article article1, Article article2) {
-        if (article1.getPrix()==article2.getPrix()){
+        if (article1.getPrix() == article2.getPrix()) {
             return article1.getNom().compareTo(article2.getNom());
         }
-        return article1.getPrix()-article2.getPrix();
+        return (int) (article1.getPrix() - article2.getPrix());
     }
 }
